pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "openrobotics"
version = "0.1.0"
description = "A comprehensive robotics development framework with MLX, LangChain, and DSPy integration"
authors = ["Nik Jois <nikjois@llamasearch.ai>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/openrobotics"
documentation = "https://openrobotics.readthedocs.io"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
mlx = "^0.1.0"
langchain = "^0.1.0"
dspy = "^0.1.0"
openai = "^1.2.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
pydantic = "^2.0.0"
sqlite-utils = "^3.35"
datasette = "^0.64.0"
llm = "^0.6.0"
numpy = "^1.24.0"
scipy = "^1.10.0"
matplotlib = "^3.7.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
typer = "^0.9.0"
rich = "^13.4.0"
PyYAML = "^6.0"

[tool.poetry.scripts]
openrobotics = "openrobotics.cli:app"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
sphinx = "^7.0.0"
sphinx-rtd-theme = "^1.2.0"
myst-parser = "^2.0.0"
pre-commit = "^3.3.2"
jupyter = "^1.0.0"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
